LPS-X Converts create_X_url utilities to TS
This conversion is pretty straightforward but showcases some interesting
things to consider:

1. Mocking the Liferay global object

Since we declare `Liferay` as a global var, mocking it is not straightforward.

  - Partial mocks are flagged as TS as incompatible
  - Reassignments in setup or teardown methods won't get picked up by TS
  forcing us to cast to `Jest.Mock` and/or other types

For the sake of progress I simply resorted to "as any" here, but maybe we
should consider a better approach. Some possible options include:
  - Simply export the type and locally cast usages of the global API
  instead of defining a global var
  - Further split the definition so we can use Partial<ILiferay> and
  Partial<IThemeDisplay> to construct partial mocks

2. String-indexed objects

We do a lot of willy-nilly object string-index access in our codebase. Maybe
as we progress our types will become more explicit, but for now it seems
like we want the type `{ [key: string]: string }` more often than not.
  - Does this type have a common name?
  - Should we export this type for general consumption or define it locally?

3. To TS or not to TS

This stems a bit from the previous point and connects with prior commits
where we addressed the fact that we keep runtime type guards in our code.

In this case, we consume `getPortletNamespace(portletID: string): string`,
but do so trying to pass `portletID: string | null`.

This hints that we might have cases of `createPortletURL` throwing exceptions
in runtime through the `getPortletNamespace` utility. Again, in the spirit of
progress, I just added a non-null assertion here `portletID!` which we know
is obviously wrong.

Updating the definition to `getPortletNamespace(portledID: string | null)`
seems to be technically correct but pushes the problem down the system. We
likely don't want a system where `null`s are being passed around unknowingly.

I'm not sure what the best approach is here, so just writing this down here
to see how this goes, specially as we move onto other modules.
