+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/global.es.js
@@ -68,10 +68,10 @@ import navigate from './util/navigate';
 import normalizeFriendlyURL from './util/normalize_friendly_url';
 import ns from './util/ns.es';
 import objectToURLSearchParams from './util/object_to_url_search_params.es';
-import createActionURL from './util/portlet_url/create_action_url.es';
-import createPortletURL from './util/portlet_url/create_portlet_url.es';
-import createRenderURL from './util/portlet_url/create_render_url.es';
-import createResourceURL from './util/portlet_url/create_resource_url.es';
+import createActionURL from './util/portlet_url/create_action_url';
+import createPortletURL from './util/portlet_url/create_portlet_url';
+import createRenderURL from './util/portlet_url/create_render_url';
+import createResourceURL from './util/portlet_url/create_resource_url';
 import {getSessionValue, setSessionValue} from './util/session.es';
 import toCharCode from './util/to_char_code.es';
 import toggleDisabled from './util/toggle_disabled';
