+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/util/portlet_url/create_render_url.ts
@@ -1,0 +1,28 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import createPortletURL from './create_portlet_url';
+
+/**
+ * Returns a render portlet URL in form of a URL object by setting the lifecycle parameter
+ */
+export default function createRenderURL(
+	basePortletURL: string,
+	parameters: {[key: string]: string} = {}
+) {
+	return createPortletURL(basePortletURL, {
+		...parameters,
+		p_p_lifecycle: '0',
+	});
+}
