+++ modules/apps/frontend-js/frontend-js-web/test/liferay/util/portlet_url/create_render_url.ts
@@ -12,24 +12,22 @@
  * details.
  */
 
-'use strict';
+import createRenderURL from '../../../../src/main/resources/META-INF/resources/liferay/util/portlet_url/create_render_url';
 
-import createResourceURL from '../../../../src/main/resources/META-INF/resources/liferay/util/portlet_url/create_resource_url.es';
-
-describe('Liferay.Util.PortletURL.createResourceURL', () => {
-	it('returns a url with the p_p_lifecycle parameter set to 2', () => {
+describe('Liferay.Util.PortletURL.createRenderURL', () => {
+	it('returns a URL object with a href parameter containing the p_p_lifecycle parameter set to 0', () => {
 		Liferay = {
 			ThemeDisplay: {
 				getPortalURL: jest.fn(() => 'http://localhost:8080'),
 			},
-		};
+		} as any;
 
-		const portletURL = createResourceURL(
+		const portletURL = createRenderURL(
 			'http://localhost:8080/group/control_panel/manage?p_p_id=foo'
 		);
 
 		expect(portletURL.href).toEqual(
-			'http://localhost:8080/group/control_panel/manage?p_p_id=foo&p_p_lifecycle=2'
+			'http://localhost:8080/group/control_panel/manage?p_p_id=foo&p_p_lifecycle=0'
 		);
 	});
 });
