+++ modules/apps/frontend-js/frontend-js-web/types/src/main/resources/META-INF/resources/index.d.ts
@@ -27,10 +27,10 @@ export { default as ItemSelectorDialog } from './liferay/ItemSelectorDialog.es';
 export { default as PortletBase } from './liferay/PortletBase.es';
 export { openModal, openSelectionModal } from './liferay/modal/Modal';
 export { default as openSimpleInputModal } from './liferay/modal/commands/OpenSimpleInputModal.es';
-export { default as createActionURL } from './liferay/util/portlet_url/create_action_url.es';
-export { default as createPortletURL } from './liferay/util/portlet_url/create_portlet_url.es';
-export { default as createRenderURL } from './liferay/util/portlet_url/create_render_url.es';
-export { default as createResourceURL } from './liferay/util/portlet_url/create_resource_url.es';
+export { default as createActionURL } from './liferay/util/portlet_url/create_action_url';
+export { default as createPortletURL } from './liferay/util/portlet_url/create_portlet_url';
+export { default as createRenderURL } from './liferay/util/portlet_url/create_render_url';
+export { default as createResourceURL } from './liferay/util/portlet_url/create_resource_url';
 export { ALIGN_POSITIONS, align, getAlignBestRegion } from './liferay/align';
 export { getAlignRegion, suggestAlignBestRegion } from './liferay/align';
 export { getSessionValue, setSessionValue } from './liferay/util/session.es';
