+++ modules/apps/frontend-js/frontend-js-web/types/src/main/resources/META-INF/resources/liferay/util/portlet_url/create_portlet_url.d.ts
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+/**
+ * Returns a portlet URL in form of a URL Object
+ */
+export default function createPortletURL(basePortletURL: string, parameters?: {
+    [key: string]: string;
+}): URL;
