+++ modules/apps/frontend-js/frontend-js-web/types/src/main/resources/META-INF/resources/liferay/util/portlet_url/create_resource_url.d.ts
@@ -11,19 +11,9 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
-
-import createPortletURL from './create_portlet_url.es';
-
 /**
  * Returns a resource portlet URL in form of a URL object by setting the lifecycle parameter
- * @param {!string} basePortletURL The base portlet URL to be modified in this utility
- * @param {object} parameters Search parameters to be added or changed in the base URL
- * @return {URL} Resource Portlet URL
- * @review
  */
-export default function createResourceURL(basePortletURL, parameters = {}) {
-	return createPortletURL(basePortletURL, {
-		...parameters,
-		p_p_lifecycle: '2',
-	});
-}
+export default function createResourceURL(basePortletURL: string, parameters?: {
+    [key: string]: string;
+}): URL;
